postgresql tables

+++ app/src/main/resources/db/migration/V2__create_product_table.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS product (
     productID SERIAL PRIMARY KEY,
     productname VARCHAR(255) NOT NULL,
     description TEXT NOT NULL,
-    price NUMERIC(10, 2) NOT NULL,
+    price FLOAT NOT NULL,
     imageUrl VARCHAR(255)
 );
